requirements_local.txt
requests
apache-airflow[celery]==2.2.3